+++ src/main/java/com/github/MaxWilmes1/ObjectOrientation/RecapProject/SecondTry/Produkt/Produkt.java
@@ -1,8 +1,10 @@
 package com.github.MaxWilmes1.ObjectOrientation.RecapProject.SecondTry.Produkt;
 
+import java.math.BigDecimal;
+
 public record Produkt(
         String id,
         String name,
-        double price
+        BigDecimal price
 ) {
 }
